// front-end/src/app/models/vehicle.ts
export class Vehicle {
  id: number;
  licencePlateNumber: string;
  brand: string;
  model: string;
  engineNumber: string;
  vinNumber: string;
  mfgDate: string;
}
export class VehicleInput {
  licencePlateNumber: string;
  brand: string;
  model: string;
  engineNumber: string;
  vinNumber: string;
  mfgDate: string;
}